Add files via upload
The objective of this project is to analyze COVID-19 data in India to gain insights into the impact of the pandemic on the country. The project aims to explore various aspects of the COVID-19 situation in India, including the number of cases, deaths, recoveries, testing rates, and trends over time.